+++ arrows/super3d/compute_depth.h
@@ -95,7 +95,7 @@ public:
     std::vector< kwiver::vital::image_container_sptr > const& frames,
     std::vector< kwiver::vital::camera_perspective_sptr > const& cameras,
     double depth_min, double depth_max,
-    unsigned int reference_frame,
+    size_t reference_frame,
     vital::bounding_box< int > const& roi,
     kwiver::vital::image_container_sptr& depth_uncertainty,
     std::vector< kwiver::vital::image_container_sptr > const& masks =
